Improve packing list sharing, ordering, and ViewModel sync with CloudKit - Loaded all packing list templates - Rebuilt category generation logic to support custom order using an array + dictionary mapping - Created dynamic categoryOrder dictionary using .enumerated() to auto-assign sort positions - Improved participant fetching from CKShare using async/await with concurrency support

+++ CoreData/Category+Extensions.swift
@@ -15,7 +15,7 @@ extension Category {
         id: UUID = UUID(),
         isExpanded: Bool = false,
         name: String,
-        position: Int,
+        position: Int64,
         items: NSSet? = nil,
         packingList: PackingList? = nil
     ) {
@@ -23,7 +23,7 @@ extension Category {
         self.id = id
         self.isExpanded = isExpanded
         self.name = name
-        self.position = Int64(position)
+        self.position = position
         self.items = items
         self.packingList = packingList
